requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
Pillow
django-filter
python-decouple
gunicorn
